Create README.md
The project focuses on addressing the challenges associated with the rapid development of neural network models, specifically the difficulty of navigating the extensive and complex solution space to identify near-optimal architectures. As new neural network models are continually being developed to tackle various datasets and tasks, each offering incremental performance improvements, finding the best model can become a time-consuming and computationally expensive process. This project aims to explore and implement efficient methods for neural architecture search (NAS) to streamline the process of discovering high-performing neural network architectures. By improving the efficiency of this search, the project seeks to reduce computational costs and accelerate the development of effective models tailored to specific applications.